+++ src/components/FcmInitializer.tsx
@@ -1,0 +1,6 @@
+import { useFcm } from "../hooks/useFcm";
+
+export default function FcmInitializer() {
+    useFcm();
+    return null;
+}